[CHORE] Repository분리
Repository분리
- Table단위로 Repository 생성하고 사용하기
