+++ src/controller/worryController.ts
@@ -3,7 +3,7 @@ import { fail, success } from "../constants/response";
 import { rm, sc } from "../constants";
 import { ClientException } from "../common/error/exceptions/customExceptions";
 import statusCode from "../constants/statusCode";
-import { worryService } from "../service";
+import { worryService, worryWithService } from "../service";
 
 const findWorryListByCategory = async (req: Request, res: Response, next: NextFunction) => {
     try {
@@ -13,7 +13,7 @@ const findWorryListByCategory = async (req: Request, res: Response, next: NextFu
             throw new ClientException("필요한 Param 값이 없습니다.");
         }
 
-        const data = await worryService.findWorryListByCategoryId(+categoryId);
+        const data = await worryWithService.findWorryListByCategoryId(+categoryId);
 
         return res.status(sc.OK).send(success(statusCode.OK, rm.READ_WORRYLIST_SUCCESS, data));
