+++ src/service/worryWithService.ts
@@ -1,7 +1,7 @@
 
 import { ClientException } from "../common/error/exceptions/customExceptions";
 import statusCode from "../constants/statusCode";
-import { withOptionRepository, worryWithRepository } from "../repository"
+import { withOptionRepository, worryWithRepository, categoryRepository } from "../repository"
 
 const chooseFinalOption = async (userId: number, worryWithId: number, optionId: number) => {
   const worryWith = await worryWithRepository.findById(worryWithId);
@@ -21,4 +21,24 @@ const chooseFinalOption = async (userId: number, worryWithId: number, optionId: 
   await worryWithRepository.updateFinalOptionById(worryWithId, optionId);
 };
 
-export default { chooseFinalOption };
+
+//~ 카테고리 별 목록조회
+const isTotal = (categoryId: number): boolean => categoryId === 0;
+
+const findWorryListByCategoryId = async (categoryId: number) => {
+  const worryWithList = isTotal(categoryId) ? await worryWithRepository.findWorries() : await worryWithRepository.findWorryListByCategoryId(categoryId);
+
+  const categoryList = await categoryRepository.getCategoryId();
+
+  if (!categoryList) {
+    throw new ClientException("카테고리가 없습니다.");
+  }
+
+  if (+categoryId > categoryList.length || +categoryId < 0) {
+    throw new ClientException("없는 카테고리입니다");
+  }
+
+  return worryWithList;
+};
+
+export default { chooseFinalOption, findWorryListByCategoryId };
